+++ MedicalCenters.Application/Features/MedicalCenter/Handlers/Commands/UpdateMedicalCenterCommandHandler.cs
@@ -19,7 +19,7 @@ namespace MedicalCenters.Application.Features.MedicalCenter.Handlers.Commands
         {
             var response = new BaseResponse();
 
-            var medicalCenter=await unitOfWork.MedicalCenterRepository.Get(command.MedicalCenterDto.Id);
+            var medicalCenter=await unitOfWork.MedicalCenterRepository.Get((int)command.MedicalCenterDto.Id);
             if (medicalCenter is null)
             {
                 string Object= "مرکز درمانی" + command.MedicalCenterDto.Id.ToString();
